Revert incorrect extended-onboarding-screen folder renaming
