+++ src/onboarding-screen/OnboardingFullScreenView.tsx
