+++ src/stacks-hierarchy/Root_ConsiderTravelTokenChangeScreen/Root_ConsiderTravelTokenChangeScreen.tsx
@@ -11,7 +11,7 @@ import {useRemoteConfig} from '@atb/RemoteConfigContext';
 
 import {useMobileTokenContextState} from '@atb/mobile-token';
 import {NoTravelTokenInfo} from './components/NoTravelTokenInfo';
-import {OnboardingFullScreenView} from '@atb/extended-onboarding-screen';
+import {OnboardingFullScreenView} from '@atb/onboarding-screen';
 import {TravelTokenBox} from '@atb/travel-token-box';
 import {useOnboardingNavigation} from '@atb/utils/use-onboarding-navigation';
 import {LoadingScreen} from '@atb/loading-screen';
