+++ src/stacks-hierarchy/Root_LocationWhenInUsePermissionScreen.tsx
@@ -4,7 +4,7 @@ import React, {useCallback} from 'react';
 import LocationWhenInUsePermissionTexts from '@atb/translations/screens/LocationWhenInUsePermission';
 import {MyLocation} from '@atb/assets/svg/color/images';
 import {useAppState} from '@atb/AppContext';
-import {OnboardingScreenComponent} from '@atb/extended-onboarding-screen';
+import {OnboardingScreenComponent} from '@atb/onboarding-screen';
 import {useGeolocationState} from '@atb/GeolocationContext';
 import {useOnboardingNavigation} from '@atb/utils/use-onboarding-navigation';
